back/requirements.txt
Django==1.8.4
wheel==0.24.0
djangorestframework==3.2.3